# Speed up minute index to session labels