--- a/setup.cfg
+++ b/setup.cfg
@@ -5,6 +5,7 @@ with-ignore-docstrings=1
 with-timer=1
 timer-top-n=15
 cover-package=trading_calendars
+logging-level=INFO
 
 [metadata]
 description-file = README.md