Instead of calling `map` of `self.minute_to_session_label`, implement a
vectorized version of `minute_index_to_session_labels` with
`direction="none"` using `searchsorted`.

On my machine, this is speeds up large lookups by a factor of over
300x.

Before:

```
In [4]: %time cal.minute_index_to_session_labels(cal.all_minutes)
CPU times: user 32.8 s, sys: 272 ms, total: 33.1 s
Wall time: 33 s
Out[4]:
DatetimeIndex(['1990-01-02', '1990-01-02', '1990-01-02', '1990-01-02',
               '1990-01-02', '1990-01-02', '1990-01-02', '1990-01-02',
               '1990-01-02', '1990-01-02',
               ...
               '2019-06-28', '2019-06-28', '2019-06-28', '2019-06-28',
               '2019-06-28', '2019-06-28', '2019-06-28', '2019-06-28',
               '2019-06-28', '2019-06-28'],
              dtype='datetime64[ns, UTC]', length=2887560, freq=None)
```

After:

```
In [4]: %time cal.minute_index_to_session_labels(cal.all_minutes)
CPU times: user 71.9 ms, sys: 23.9 ms, total: 95.8 ms
Wall time: 95.1 ms
Out[4]:
DatetimeIndex(['1990-01-02', '1990-01-02', '1990-01-02', '1990-01-02',
               '1990-01-02', '1990-01-02', '1990-01-02', '1990-01-02',
               '1990-01-02', '1990-01-02',
               ...
               '2019-06-28', '2019-06-28', '2019-06-28', '2019-06-28',
               '2019-06-28', '2019-06-28', '2019-06-28', '2019-06-28',
               '2019-06-28', '2019-06-28'],
              dtype='datetime64[ns, UTC]', length=2887560, freq=None)
```